feat: Create convert_currency endpoint and add helpers for format code

+++ lib/core/lib/currency_convertor/commands/convert_currency.ex
@@ -14,7 +14,7 @@ defmodule Core.CurrencyConvertor.Commands.ConvertCurrency do
 
   embedded_schema do
     field :final_currency, :string
-    field :value, :integer
+    field :value, :float
     field :convertion_tax, :float
     field :converted_value, :string
   end
